highscore.json nun im Hauptverzeichnis
