+++ .gitignore
@@ -9,4 +9,4 @@ Thumbs.db
 *.swp
 *.swo
 *~
-Breakout/highscores.json
+highscores.json
