--- Breakout/highscores.json
@@ -1,24 +1,0 @@
-{
-  "Breakout": [
-    {
-      "name": "SON",
-      "score": 58
-    },
-    {
-      "name": "SON",
-      "score": 48
-    },
-    {
-      "name": "SON",
-      "score": 43
-    },
-    {
-      "name": "SON",
-      "score": 42
-    },
-    {
-      "name": "SON",
-      "score": 38
-    }
-  ]
-}
