--- Snake/highscores.json
@@ -1,24 +1,0 @@
-{
-  "Snake": [
-    {
-      "name": "SON",
-      "score": 40
-    },
-    {
-      "name": "SON",
-      "score": 40
-    },
-    {
-      "name": "SON",
-      "score": 40
-    },
-    {
-      "name": "SON",
-      "score": 20
-    },
-    {
-      "name": "SON",
-      "score": 20
-    }
-  ]
-}
